Added x64 build configuration and fixed 64-bit compilation.

+++ ResourceString.cpp
@@ -26,7 +26,7 @@ tstring loadString(HMODULE module, uint resourceId)
 
 	for (;;)
 	{
-		const int bufsize = length+1;
+		const int bufsize = static_cast<int>(length+1);
 		const int copied  = ::LoadString(module, resourceId, buffer.get(), bufsize);
 
 		if (copied == 0)
